+++ CHANGELOG.md
@@ -5,9 +5,22 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ---
 
+[1.1.0]: https://github.com/qbasic16/swiss_uid/releases/tag/1.1.0
 [1.0.2]: https://github.com/qbasic16/swiss_uid/releases/tag/1.0.2
 [1.0.1]: https://github.com/qbasic16/swiss_uid/releases/tag/1.0.1
 [1.0.0]: https://github.com/qbasic16/swiss_uid/releases/tag/1.0.0
+
+## [1.1.0] - 2025-01-26
+
+### Breaking
+
+- Changed method signature and body of `SwissUid::checkdigit` to return the
+  contained checkdigit instead of recalculating it
+
+### Changed
+
+- Refactored nibble utils
+- Cleaned up code
 
 ## [1.0.2] - 2024-07-07